compact module structure
